Turn off HID endpoint to allow more USB devices
The HID endpoint reserves a fixed amount of bandwidth when it is
turned on.  This limits the number of boards that can be connected
to a single USB hub to between 9 and 19 boards.  This patch turns
off the HID endpoint so a large number of devices can be connected
to USB.